purchase_tracker_ds/code/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.1.0
whitenoise==6.6.0
channels==4.1.0
daphne==4.1.2
paho-mqtt==2.1.0
pyzmq==26.0.3 
requests==2.31.0
python-dateutil==2.8.2

